Make listener retention ticks long

+++ common/src/main/java/me/wyne/wutils/common/event/RegisterableListener.java
@@ -1,7 +1,7 @@
 package me.wyne.wutils.common.event;
 
 public interface RegisterableListener {
-    default int getRetentionTicks() {
+    default long getRetentionTicks() {
         if (getClass().isAnnotationPresent(ListenerRetention.class)) {
             return getClass().getDeclaredAnnotation(ListenerRetention.class).ticks();
         }
